Rearrange files and install pytest cov
