+++ snakey/game.py
@@ -1,10 +1,9 @@
-import argparse
 from time import time
 from typing import List
 
-from . import colors as _colors
-from .pair import *
-from .languages import LANGUAGES
+from snakey import colors as _colors
+from snakey.pair import *
+from snakey.languages import LANGUAGES
 import tkinter as tk
 
 
@@ -1018,16 +1017,3 @@ class SnaKeyGUI(tk.Tk):
         self.game.chaser_tile().color(cs['chaser'])
         self.game.nommer_tile().color(cs['nommer'])
         self.game.runner_tile().color(cs['runner'])
-
-
-def parse_cli_arguments():
-    parser = argparse.ArgumentParser()
-    parser.add_argument('-w', '--grid-width',
-                        action='store', type=int, required=False, default=10,
-                        help='Width and height of the square grid')
-    return parser.parse_args()
-
-if __name__ == '__main__':
-    arguments = parse_cli_arguments()
-    root = SnaKeyGUI(arguments.grid_width)
-    root.mainloop()
