+++ tox.ini
@@ -10,5 +10,6 @@ skipsdist=True
 [testenv]
 deps =
     pytest
+    pytest-cov
 commands =
-    pytest -v -rf {posargs:tests}
+    pytest -v --cov=snakey --cov-report html:cov_html -rf {posargs:tests}
